\docType{data}
\name{ninames}
\alias{ninames}
\title{Name and Gender Data for Northern Ireland}
\format{A dataframe with 12806 observations on the following 4 variables,

\describe{

\item{Name}{String: Birth name as recorded by NISRA, may be changed from uppercase.}

\item{F}{Numeric: Counts for births with that name recorded as female in a Year}

\item{M}{Numeric: Counts for births with that name recorded as male in a Year}

\item{Year}{Numeric: Years associated with given counts for a name}

}}
\source{
  http://www.nisra.gov.uk/demography/default.asp28.htm
}
\description{
  The Northern Ireland Statistics and Research
  Administration provides gender data for names with
  greater than 3 births per year. Full name data is
  provided between 1997 and 2011.
}
\references{
  Northern Ireland Statistics and Research Administration,
  Baby Names Statistical Bulletin [computer file].
  Demography Branch
}
\keyword{datasets}

