\name{readNISRANames}
\alias{readNISRANames}
\title{Read and return Northern Ireland Name Data}
\usage{
  readNISRANames()
}
\value{
  Data frame with columns for Name, Year, and counts for
  gender incidence
}
\description{
  Read and return Northern Ireland Name Data
}
\seealso{
  \code{\link{readONSNames}},
  \code{\link{readScotlandNames}},
  \code{\link{readSSANames}}
}
\keyword{ireland}

