\name{readScotlandNames}
\alias{readScotlandNames}
\title{Read and return Scotland name data}
\usage{
  readScotlandNames()
}
\value{
  Data frame with columns for Name, Year, and counts for
  gender incidence
}
\description{
  Download data from the Scotland GRO website and convert
  into a single data frame
}
\seealso{
  \code{\link{readONSNames}}, \code{\link{readNISRANames}},
  \code{\link{readSSANames}}
}
\keyword{scotland}

