\docType{data}
\name{scotnames}
\alias{scotnames}
\title{Name and Gender Data for Scotland}
\format{A dataframe with 13605 observations on the following 4 variables,

\describe{

\item{Name}{String: Birth name as recorded by the GRO, may be changed from uppercase.}

\item{F}{Numeric: Counts for births with that name recorded as female in a Year}

\item{M}{Numeric: Counts for births with that name recorded as male in a Year}

\item{Year}{Numeric: Years associated with given counts for a name}

}}
\source{
  http://www.gro-scotland.gov.uk/statistics/theme/vital-events/births/bckgr-info.html
}
\description{
  Scotland's General Register Office provides gender data
  for names with greater than 3 births per year. Full name
  data is provided for 2009 and 2010.
}
\references{
  General Register Office, Baby Names Statistical Bulletin
  [computer file]. National Records of Scotland
}
\keyword{datasets}

