\docType{data}
\name{uknames}
\alias{uknames}
\title{Name and Gender Data for the United Kingdom}
\format{A dataframe with 171695 observations on the following 4 variables,

\describe{

\item{Name}{String: Birth name as recorded by agencies, may be changed from uppercase.}

\item{F}{Numeric: Counts for births with that name recorded as female in a Year}

\item{M}{Numeric: Counts for births with that name recorded as male in a Year}

\item{Year}{Numeric: Years associated with given counts for a name}

}}
\description{
  Merged data from England and Wales, Northern Ireland and
  Scotland
}
\keyword{datasets}

