\docType{data}
\name{ukprocessed}
\alias{ukprocessed}
\title{Processed name data for the United Kingdom}
\format{A dataframe with 30714 observations on the following 7 variables,

\describe{

\item{Name}{String: Birth name as recorded by agencies, may be changed from uppercase}

\item{YearsAppearing}{Numeric: Number of years where the name appears}

\item{CountsFemale}{Numeric: Counts for births with that name recorded as female}

\item{CountsMale}{Numeric: Counts for births with that name recorded as male}

\item{ProbGender}{Factor: Likely gender, outputted from the default classifier}

\item{Upper}{Numeric: Upper bound for confidence interval}

\item{Lower}{Numeric: Lower bound for confidence interval}

}}
\description{
  Processed data from England and Wales, Northern Ireland
  and Scotland
}
\keyword{datasets}

