\name{byNameCount}
\alias{byNameCount}
\title{Read name data formatted as gender-year combinations and
  accumulates total counts per name}
\usage{
  byNameCount(data)
}
\arguments{
  \item{data}{A data frame with columns for Name, F, M, and
  Year}
}
\value{
  Data frame with columns for Name, counts for gender
  incidence and year counts
}
\description{
  Read name data formatted as gender-year combinations and
  accumulates total counts per name
}

