\name{readNISRANames}
\alias{readNISRANames}
\title{Read and return Northern Ireland Name Data}
\usage{
  readNISRANames()
}
\value{
  Data frame with columns for Name, Year, and counts for
  gender incidence
}
\description{
  Download data from the NISRA website and convert into a
  single data frame Downloading and converting data will
  take some time. The resultant dataset is provided as
  \code{\link{ninames}}
}
\seealso{
  \code{\link{readONSNames}},
  \code{\link{readScotlandNames}},
  \code{\link{readSSANames}}
}
\keyword{ireland}

