\name{readONSNames}
\alias{readONSNames}
\title{Read and return name data for England and Wales from the
  Office of National Statistics}
\usage{
  readONSNames()
}
\value{
  Data frame with columns for Name, Year, and counts for
  gender incidence
}
\description{
  Read and return name data for England and Wales from the
  Office of National Statistics
}
\seealso{
  \code{\link{readScotlandNames}},
  \code{\link{readNISRANames}}, \code{\link{readSSANames}}
}
\keyword{england}

