\name{readSSANames}
\alias{readSSANames}
\title{Read and return United States name data from the
  Social Security Administration}
\usage{
  readSSANames()
}
\value{
  Data frame with columns for Name, Year, and counts for
  gender incidence
}
\description{
  Read and return United States name data from the Social
  Security Administration
}
\seealso{
  \code{\link{readONSNames}}, \code{\link{readNISRANames}},
  \code{\link{readScotlandNames}}
}
\keyword{america}

